move SparkXXCatalog
