+++ spark-connector/spark-connector/src/main/java/com/datastrato/gravitino/spark/connector/plugin/GravitinoDriverPlugin.java
@@ -8,8 +8,8 @@ package com.datastrato.gravitino.spark.connector.plugin;
 import com.datastrato.gravitino.Catalog;
 import com.datastrato.gravitino.spark.connector.GravitinoSparkConfig;
 import com.datastrato.gravitino.spark.connector.catalog.GravitinoCatalogManager;
-import com.datastrato.gravitino.spark.connector.catalog.GravitinoHiveCatalog;
-import com.datastrato.gravitino.spark.connector.catalog.GravitinoIcebergCatalog;
+import com.datastrato.gravitino.spark.connector.hive.GravitinoHiveCatalog;
+import com.datastrato.gravitino.spark.connector.iceberg.GravitinoIcebergCatalog;
 import com.google.common.base.Preconditions;
 import java.util.Collections;
 import java.util.Locale;
